TS-migration: ListItemIcon
