+++ src/components/Avatar/Avatar.tsx
@@ -10,7 +10,7 @@ import { AvatarContent } from "./AvatarContent";
 import Tooltip from "../Tooltip/Tooltip";
 import ClickableWrapper from "../Clickable/ClickableWrapper";
 import Dialog from "../Dialog/Dialog";
-import { iconSubComponentProps } from "../Icon/Icon";
+import { IconSubComponentProps } from "../Icon/Icon";
 
 import "./Avatar.scss";
 import VibeComponentProps from "src/types/VibeComponentProps";
@@ -26,7 +26,7 @@ export interface AvatarProps extends VibeComponentProps {
   tooltipProps?: any;
   ariaLabel?: string;
   withoutTooltip?: boolean;
-  icon?: string | React.FunctionComponent<iconSubComponentProps> | null;
+  icon?: string | React.FunctionComponent<IconSubComponentProps> | null;
   type?: AvatarTypes;
   className?: string;
   textClassName?: string;
