+++ src/components/Icon/Icon.tsx
@@ -1,4 +1,4 @@
-import React, { forwardRef, Ref, CSSProperties } from "react";
+import React, { CSSProperties, forwardRef, Ref } from "react";
 import cx from "classnames";
 import useMergeRefs from "../../hooks/useMergeRefs";
 import { IconType } from "./IconConstants";
@@ -11,9 +11,9 @@ import "./Icon.scss";
 import { MouseEventCallBack } from "src/types/events";
 
 // eslint-disable-next-line no-unused-vars,@typescript-eslint/no-unused-vars,@typescript-eslint/no-empty-function
-const CLICK_NOOP = (event: React.MouseEvent) => {};
+const CLICK_NOOP = (_event: React.MouseEvent) => {};
 
-export interface iconSubComponentProps {
+export interface IconSubComponentProps {
   ref?: Ref<HTMLElement>;
   id?: string;
   size?: string | number;
@@ -26,7 +26,7 @@ export interface iconSubComponentProps {
 interface IconProps extends VibeComponentProps {
   // eslint-disable-next-line no-unused-vars
   onClick?: (event: React.MouseEvent) => void;
-  icon: string | React.FunctionComponent<iconSubComponentProps> | null;
+  icon: string | React.FunctionComponent<IconSubComponentProps> | null;
   clickable?: boolean;
   iconLabel?: string;
   iconType?: IconType;
