+++ src/components/ListItemIcon/ListItemIconConstants.ts
@@ -1,0 +1,4 @@
+export enum ListItemIconMargin {
+  START = "start",
+  END = "end"
+}
