+++ src/components/ListItemIcon/__stories__/ListItemIcon.stories.mdx
@@ -1,10 +1,18 @@
 import ListItemIcon from "../ListItemIcon";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate } from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+
+export const metaSettings = createStoryMetaSettings({
+  component: ListItemIcon,
+  enumPropNamesArray: ["margin"],
+  iconPropNamesArray: ["icon"]
+});
 
 <Meta
   title="Navigation/List/ListItemIcon"
-  component={ ListItemIcon }
+  component={ListItemIcon}
+  argTypes={metaSettings.argTypes}
+  decorators={metaSettings.decorators}
 />
 
 <!--- Component template -->
@@ -14,6 +22,7 @@ export const listItemIconTemplate = createComponentTemplate(ListItemIcon);
 <!--- Component documentation -->
 
 # ListItemIcon
+
 - [Overview](#overview)
 - [Props](#props)
 - [Usage](#usage)
@@ -24,14 +33,13 @@ export const listItemIconTemplate = createComponentTemplate(ListItemIcon);
 - [Feedback](#feedback)
 
 ## Overview
+
 Icon of a list item.
 
 <Canvas>
-  <Story name="Overview"
-         args={ {prop1: "prop's default value", prop2: "prop's default value"} }>
-    { listItemIconTemplate.bind({}) }
-  </Story>
+  <Story name="Overview">{listItemIconTemplate.bind({})}</Story>
 </Canvas>
 
 ## Props
-<ArgsTable of={ ListItemIcon } />
+
+<ArgsTable story="Overview" />
